pyproject.toml
[tool.poetry]
name = "djangoprojectdrf-1-0"
version = "0.1.0"
description = ""
authors = ["ryfe <denis.koptelev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
black = "^24.4.2"
django-filter = "^24.2"
ipython = "^8.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
